Code changes after moving featureRelation, sensorSystem and compositePhenomenon to core mapping
