+++ hibernate/common/src/main/java/org/n52/sos/ds/hibernate/dao/FeatureOfInterestDAO.java
@@ -49,7 +49,6 @@ import org.n52.sos.ds.hibernate.entities.FeatureOfInterestType;
 import org.n52.sos.ds.hibernate.entities.ObservationConstellation;
 import org.n52.sos.ds.hibernate.entities.Offering;
 import org.n52.sos.ds.hibernate.entities.RelatedFeature;
-import org.n52.sos.ds.hibernate.entities.TFeatureOfInterest;
 import org.n52.sos.ds.hibernate.entities.observation.AbstractObservation;
 import org.n52.sos.ds.hibernate.entities.observation.legacy.ContextualReferencedLegacyObservation;
 import org.n52.sos.ds.hibernate.entities.observation.series.ContextualReferencedSeriesObservation;
@@ -238,13 +237,8 @@ public class FeatureOfInterestDAO extends AbstractIdentifierNameDescriptionDAO i
         Criteria criteria = session.createCriteria(FeatureOfInterest.class);
         ProjectionList projectionList = Projections.projectionList();
         projectionList.add(Projections.property(FeatureOfInterest.IDENTIFIER));
-
-        //get parents if transactional profile is active
-        boolean tFoiSupported = HibernateHelper.isEntitySupported(TFeatureOfInterest.class);
-        if (tFoiSupported) {
-            criteria.createAlias(TFeatureOfInterest.PARENTS, "pfoi", JoinType.LEFT_OUTER_JOIN);
-            projectionList.add(Projections.property("pfoi." + FeatureOfInterest.IDENTIFIER));
-        }
+        criteria.createAlias(FeatureOfInterest.PARENTS, "pfoi", JoinType.LEFT_OUTER_JOIN);
+        projectionList.add(Projections.property("pfoi." + FeatureOfInterest.IDENTIFIER));
         criteria.setProjection(projectionList);
         //return as List<Object[]> even if there's only one column for consistency
         criteria.setResultTransformer(NoopTransformerAdapter.INSTANCE);
@@ -256,9 +250,7 @@ public class FeatureOfInterestDAO extends AbstractIdentifierNameDescriptionDAO i
         for(Object[] result : results) {
             String featureIdentifier = (String) result[0];
             String parentFeatureIdentifier = null;
-            if (tFoiSupported) {
                 parentFeatureIdentifier = (String) result[1];
-            }
             if (parentFeatureIdentifier != null) {
                 CollectionHelper.addToCollectionMap(featureIdentifier, parentFeatureIdentifier, foiMap);
             } else {
@@ -299,7 +291,7 @@ public class FeatureOfInterestDAO extends AbstractIdentifierNameDescriptionDAO i
             final Session session) {
         FeatureOfInterest feature = getFeatureOfInterest(identifier, session);
         if (feature == null) {
-            feature = new TFeatureOfInterest();
+            feature = new FeatureOfInterest();
             feature.setIdentifier(identifier);
             if (url != null && !url.isEmpty()) {
                 feature.setUrl(url);
@@ -326,7 +318,7 @@ public class FeatureOfInterestDAO extends AbstractIdentifierNameDescriptionDAO i
      * @param session
      *            Hibernate session
      */
-    public void insertFeatureOfInterestRelationShip(final TFeatureOfInterest parentFeature,
+    public void insertFeatureOfInterestRelationShip(final FeatureOfInterest parentFeature,
             final FeatureOfInterest childFeature, final Session session) {
         parentFeature.getChilds().add(childFeature);
         session.saveOrUpdate(parentFeature);
@@ -351,7 +343,7 @@ public class FeatureOfInterestDAO extends AbstractIdentifierNameDescriptionDAO i
         if (CollectionHelper.isNotEmpty(relatedFeatures)) {
             for (final RelatedFeature relatedFeature : relatedFeatures) {
             	if (!featureOfInterest.getIdentifier().equals(relatedFeature.getFeatureOfInterest().getIdentifier())) {
-	                insertFeatureOfInterestRelationShip((TFeatureOfInterest) relatedFeature.getFeatureOfInterest(),
+	                insertFeatureOfInterestRelationShip(relatedFeature.getFeatureOfInterest(),
 	                        featureOfInterest, session);
             	}
             }
