+++ hibernate/common/src/main/java/org/n52/sos/ds/hibernate/entities/TFeatureOfInterest.java
@@ -28,64 +28,14 @@
  */
 package org.n52.sos.ds.hibernate.entities;
 
-import java.util.HashSet;
-import java.util.Set;
-
-import org.n52.sos.ds.hibernate.entities.HibernateRelations.HasParentChilds;
-
-import com.google.common.collect.Sets;
-
 /**
  * @author <a href="mailto:c.hollmann@52north.org">Carsten Hollmann</a>
  *
  * @since 4.0.0
  */
-public class TFeatureOfInterest extends FeatureOfInterest implements
-        HasParentChilds<FeatureOfInterest> {
+@Deprecated
+public class TFeatureOfInterest extends FeatureOfInterest {
     private static final long serialVersionUID = -880472749711995015L;
 
-    private Set<FeatureOfInterest> childs = Sets.newHashSet();
-    private Set<FeatureOfInterest> parents = Sets.newHashSet();
-
-    @Override
-    public Set<FeatureOfInterest> getParents() {
-        return parents;
-    }
-
-    @Override
-    public void setParents(final Set<FeatureOfInterest> parents) {
-        this.parents = parents;
-    }
-
-    @Override
-    public Set<FeatureOfInterest> getChilds() {
-        return childs;
-    }
-
-    @Override
-    public void setChilds(final Set<FeatureOfInterest> childs) {
-        this.childs = childs;
-    }
-
-    @Override
-    public void addParent(FeatureOfInterest parent) {
-        if (parent == null) {
-            return;
-        }
-        if (this.parents == null) {
-            this.parents = new HashSet<>();
-        }
-        this.parents.add(parent);
-    }
-
-    @Override
-    public void addChild(FeatureOfInterest child) {
-        if (child == null) {
-            return;
-        }
-        if (this.childs == null) {
-            this.childs = new HashSet<>();
-        }
-        this.childs.add(child);
-    }
+    
 }
